requirements.txt
rembg
pytesseract
discord.py
tinydb
pytest
black

